+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Version Next
 
+- *Breaking:* Classes in `recipemd.data` are now frozen
+- Add option `--export-linked` to export linked recipes in the correct scale to a folder 
 - Amount values can be negative to allow simple stock keeping
 - Remove linked recipe amount from title in instructions as it is confusing with multiple levels of flattened recipes
 - Prevent duplicated headlines when flattening